Moved texture stuff to own module
Closes https://github.com/PistonDevelopers/hematite/issues/18
